Refactor dataloader.py
